Fri, Sep 27, 2024, 10:41 AM -04:00

+++ readability/readability.c
@@ -16,20 +16,20 @@ int main (void)
 {
     string text = get_input("Text");
 
-    int no_of_spaces = count(text, 32);
+    float no_of_spaces = count(text, 32);
     if (no_of_spaces <= 100) {
 
         // total letters divided by total words
-        float L = ( (float)count_alpha(text) / ++no_of_spaces ) * 100;
+        float L = ( count_alpha(text) / ++no_of_spaces ) * 100;
         printf("%f\n", L);
 
         // count punctuation to find number of sentences
         int total_sentences = count(text, 33) + count(text, 63) + count(text, 46);
-        int S = ( total_sentences / ++no_of_spaces ) * 100;
-        printf("%i\n", S);
+        float S = ( total_sentences / ++no_of_spaces ) * 100;
+        printf("%f\n", S);
 
-        int result = (0.0588 * L) - (0.296 * S) - 15.8;
-        printf("%i\n", result);
+        float result = (0.0588 * L) - (0.296 * S) - 15.8;
+        printf("%f\n", result);
 
     }
     else
